% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotenuse.R
\name{hypotenuse}
\alias{hypotenuse}
\title{Length of the hypotenuse}
\usage{
hypotenuse(a, b)
}
\arguments{
\item{a}{adjacent side of a right triangle}

\item{b}{opposite side of a triangle}
}
\value{
length of hypotenuse
}
\description{
Length of the hypotenuse
}
\details{
Based on Pythagorean Theorem

\deqn{a^2+b^2=c^2}
}
