% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotenuse.R
\name{hypotenuse}
\alias{hypotenuse}
\title{calculates the hypo.}
\usage{
hypotenuse(a, b)
}
\arguments{
\item{a}{side one}

\item{b}{side two}
}
\value{
The length of the third side.
}
\description{
calculates the hypo.
}
